Add a new ResultHandle type for the result of SendNotify
The handle contains the channel that was formerly returned from
SendNotify, but it also supports a Close method so that we can stop
waiting for the result of the command, and free up the tracking entry in
the client.

This is necessary for nCLI on TGXL (which has some commands that it
honors but doesn't return any response to), but it's also a generally
nice API improvement.