parse meter packets in-library and pass them to a channel
this code used to be in nCAT, but only because I couldn't figure out at
the time what the interface should be. Naturally other apps will want to
be able to get meter updates as well.

Now you can call `SetMeterChan` to get fully-parsed `MeterReport`s on
the passed channel. Flexclient will handle the mapping between meter IDs
and name/source for you, and it will attempt to decode the raw int16
value into a float according to the unit in the metadata (the raw value
is also available).

You still need to send "sub meter all" to the radio to have it send
meter metadata and meter packets, and you need to `RunUDP` to receive
the meter packets (or any VITA at all).

Meter updates are available as regular state updates for clients that
need to know what meters are available.